+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - Replace `crc32c` dependency with `crc-any`
+
 ## [0.11.6] - 2024-02-06
 
 ### Added